ES6 support
Webpack was added to this project (npm run build) but it can only be
used after changing all files to make use of modules (require/import
and export/exports) because Webpack uses import/require statements to
navigate through the code and create the dependency tree. We can write
ES6 code even before changing the entire code but in this case we still
have to use Grunt.